requirements.txt
dj-database-url==0.5.0
Django==2.1
django-cors-headers==2.1.0
django-filter==2.0.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2==2.8.3
PyJWT==1.5.3
pytz==2018.5
whitenoise==4.1.3
